Created Jpa repository interfaces. Added method to deleteById to LinkService interface.

+++ scrapper/src/main/java/ru/tinkoff/edu/java/scrapper/repository/pojo/Chat.java
@@ -13,12 +13,12 @@ import lombok.Setter;
 public class Chat {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private long id;
 
     public Chat() {
     }
 
-    public Chat(int id) {
+    public Chat(long id) {
         this.id = id;
     }
 }
